requirements.txt
streamlit==1.28.1
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.0
pandas>=2.2.0
python-dotenv==1.0.0
rapidfuzz==3.5.2
lxml==4.9.3